// lib/api/model/reviewableLineItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * API Documentation
 * The API description
 * OpenAPI spec version: 1.0
 */
import type { ReviewableLineItemVariant } from './reviewableLineItemVariant';

export interface ReviewableLineItem {
  lineItemId: string;
  productId: string;
  title: string;
  quantity: number;
  orderId: string;
  orderNumber: number;
  processedAt: string;
  variant?: ReviewableLineItemVariant;
  hasReviewed: boolean;
}